# Refactor(pdf generation): replace quicktype, add static getter of pdf parameters