--- a/app/services/pdf/beratungshilfe/sections/F_besitz/__test__/F_besitz.test.ts
+++ b/app/services/pdf/beratungshilfe/sections/F_besitz/__test__/F_besitz.test.ts
@@ -4,7 +4,7 @@
 /* eslint-disable sonarjs/no-duplicate-string */
 
 import { type BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
-import { getBeratungshilfeParameters } from "data/pdf/beratungshilfe/beratungshilfe.generated";
+import { getBeratungshilfeParameters } from "~/services/pdf/beratungshilfe/beratungshilfe.server";
 import {
   createAttachment,
   newPageHint,