--- a/app/services/pdf/beratungshilfe/sections/__test__/A_angelegenheit.test.ts
+++ b/app/services/pdf/beratungshilfe/sections/__test__/A_angelegenheit.test.ts
@@ -7,7 +7,7 @@ import {
   createAttachment,
   newPageHint,
 } from "~/services/pdf/beratungshilfe/attachment";
-import { getBeratungshilfeParameters } from "~/services/pdf/beratungshilfe/beratungshilfe.server";
+import { getBeratungshilfeParameters } from "data/pdf/beratungshilfe/beratungshilfe.generated";
 import {
   fillAngelegenheit,
   BESCHREIBUNG_ANGELEGENHEIT_TITLE,
@@ -27,7 +27,7 @@ describe("A_angelegenheit", () => {
       eigeninitiativeBeschreibung: "eigeninitiativeBeschreibung",
     };
     const attachment = createAttachment(context);
-    const pdfFields = await getBeratungshilfeParameters();
+    const pdfFields = getBeratungshilfeParameters();
 
     fillAngelegenheit(attachment, pdfFields, context);
 
@@ -57,7 +57,7 @@ describe("A_angelegenheit", () => {
         "eigeninitiativeBeschreibung eigeninitiativeBeschreibung eigeninitiativeBeschreibung eigeninitiativeBeschreibung",
     };
     const attachment = createAttachment(context);
-    const pdfFields = await getBeratungshilfeParameters();
+    const pdfFields = getBeratungshilfeParameters();
 
     fillAngelegenheit(attachment, pdfFields, context);