--- a/app/services/pdf/beratungshilfe/sections/__test__/A_angelegenheit.test.ts
+++ b/app/services/pdf/beratungshilfe/sections/__test__/A_angelegenheit.test.ts
@@ -7,7 +7,7 @@ import {
   createAttachment,
   newPageHint,
 } from "~/services/pdf/beratungshilfe/attachment";
-import { getBeratungshilfeParameters } from "data/pdf/beratungshilfe/beratungshilfe.generated";
+import { getBeratungshilfeParameters } from "~/services/pdf/beratungshilfe/beratungshilfe.server";
 import {
   fillAngelegenheit,
   BESCHREIBUNG_ANGELEGENHEIT_TITLE,