--- a/app/services/pdf/fileTypes.ts
+++ b/app/services/pdf/fileTypes.ts
@@ -12,7 +12,3 @@ type JsonField = BooleanField | StringField;
 export function isBooleanField(field: JsonField): field is BooleanField {
   return typeof field.value === "boolean";
 }
-
-export function isStringField(field: JsonField): field is StringField {
-  return typeof field.value === "string";
-}