--- a/app/services/pdf/pdf.server.ts
+++ b/app/services/pdf/pdf.server.ts
@@ -1,6 +1,4 @@
 import type { PDFForm } from "pdf-lib";
-import { PDFCheckBox, PDFTextField } from "pdf-lib";
-
 import type { BooleanField, StringField } from "./fileTypes";
 import { getFontSizeFieldValue } from "./getFontSizeFieldValue";
 import { lowercaseFirstLetter } from "~/util/strings";
@@ -27,21 +25,16 @@ export function normalizePropertyName(propertyName: string) {
 
 export function changeBooleanField(booleanField: BooleanField, form: PDFForm) {
   if (!booleanField.value) return;
-  const formField = form.getField(booleanField.name);
-  if (formField instanceof PDFCheckBox) {
-    formField.check();
-  }
+  form.getCheckBox(booleanField.name).check();
 }
 
 export function changeStringField(stringField: StringField, form: PDFForm) {
   if (!stringField.value) return;
-  const formField = form.getField(stringField.name);
-  if (formField instanceof PDFTextField) {
-    // We override the restriction of a field where it is set to have max length of 2.
-    // This is because it impacts letter spacing e.g. in the fourth column of "G_ausgaben".
-    formField.setMaxLength(undefined);
-    formField.setText(stringField.value);
-    const fontSize = getFontSizeFieldValue(stringField.name);
-    formField.setFontSize(fontSize);
-  }
+  const formField = form.getTextField(stringField.name);
+  // We override the restriction of a field where it is set to have max length of 2.
+  // This is because it impacts letter spacing e.g. in the fourth column of "G_ausgaben".
+  formField.setMaxLength(undefined);
+  formField.setText(stringField.value);
+  const fontSize = getFontSizeFieldValue(stringField.name);
+  formField.setFontSize(fontSize);
 }