… app/services/pdf/beratungshilfe/sections/__test__/B_vorraussetzungen.test.ts
… app/services/pdf/beratungshilfe/sections/__test__/C_einkommen.test.ts
… app/services/pdf/beratungshilfe/sections/__test__/D_wohnen.test.ts
… app/services/pdf/beratungshilfe/sections/__test__/G_ausgaben.test.ts
… app/services/pdf/beratungshilfe/sections/__test__/header.test.ts
… data/pdf/beratungshilfe/beratungshilfe.generated.ts