- removes quicktype dependency, instead export our own types
- also export static `getBeratungshilfePdfFromContext()`, which cleanly returns a new object to fill PDFs with (no await, no need to parse pdf etc)

@pgurusinga I had to adapt a few pdf unit tests, mostly where the tests previously relied on `false` instead of `undefined`. I think the new behaviour is correct but please double-check 🙏🏼 

Also, I think the unit tests are now much quicker (4-5 s on my machine, instead of ~10s)

some more stats on the /pdf endpoint (using 10s autocannon):
- before: avg latency: 468 ms, avg req/sec: 21
- after: avg latency: 352 ms, avg req/sec: 28