@pgurusinga 2024-05-30
@chohner I'm missing some context. I noticed that the getBeratungshilfeParameters function returns a mapping of field names to structured names. Do I need to manually add the mapping for future documents?
 
![image](https://github.com/digitalservicebund/a2j-rechtsantragstelle/assets/9930966/2de1ae1f-2253-4316-9b7d-bd7d8343beed)